requirements.txt
langchain
langchain-community
langchain-huggingface
faiss-cpu
python-dotenv
torch
streamlit
sentence-transformers